+++ python/ray/experimental/workflow/tests/test_signature_check.py
@@ -1,6 +1,6 @@
 import pytest
 
-import ray
+from ray.tests.conftest import *  # noqa
 from ray.experimental import workflow
 
 
@@ -9,9 +9,11 @@ def signature_check(a, b, c=1):
     pass
 
 
-def test_signature_check():
-    ray.init(namespace="workflow")
-
+@pytest.mark.parametrize(
+    "ray_start_regular", [{
+        "namespace": "workflow"
+    }], indirect=True)
+def test_signature_check(ray_start_regular):
     with pytest.raises(TypeError):
         signature_check(1, 2)
 
@@ -31,5 +33,3 @@ def test_signature_check():
     signature_check.step(1, 2, c=3)
     signature_check.step(1, b=2, c=3)
     signature_check.step(a=1, b=2, c=3)
-
-    ray.shutdown()
